refactor bpf program file structure
